pyproject.toml
# FILE: pyproject.toml
[project]
name = "maiming"
version = "2.0.0"
requires-python = ">=3.11"
dependencies = ["PyQt6>=6.6", "numpy>=1.26", "PyOpenGL>=3.1.7"]

[tool.ruff]
line-length = 100